rsa_authorize/requirements0722.txt
pycryptodome == 3.15.0
python-dotenv == 0.20.0
fastapi == 0.79.0
uvicorn == 0.18.2
pycryptodomex == 3.15.0
SQLAlchemy == 1.4.39
passlib == 1.7.4
python-jose == 3.3.0
python-multipart == 0.0.5

autopep8 == 1.6.0